backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
pydantic>=2.5.0
python-dotenv>=1.0.0
akshare>=1.12.0
vectorbt>=0.26.0
apscheduler>=3.10.4
httpx>=0.25.0
